fix iOS resolution issue (#3066)

+++ native/cocos/platform/Device.h
@@ -77,7 +77,7 @@ public:
     /**
      * Get device pixel ratio.
      */
-    static int getDevicePixelRatio();
+    static float getDevicePixelRatio();
     
     /**
      * To enable or disable accelerometer.
